Delete ALert

+++ src/Components/pages/AdminPage/ManageChapter/DeleteChapter.jsx
@@ -31,7 +31,9 @@ function DeleteChapter() {
 
     const handleSubmit = (e) => {
         e.preventDefault()
-        if (formValue.course && formValue.part && formValue.chapter) {
+        var deleteConfirm = prompt(`Please Type 'DELETE' in the field below.`, "");
+        if (formValue.course && formValue.part && formValue.chapter &&
+            deleteConfirm != null && deleteConfirm == "DELETE") {
             setSectionLoading(true)
             axios.delete(`${process.env.REACT_APP_API_DOMAIN}/chapter/delete/${formValue.chapter}`, {
                 headers: {
